# DO NOT MERGE: test potential MustApplyAutoAliases speedup