Exploring by commenting out the bits that don't look like they are only needed during gen. Get about a 2x speed up.

before:
```
❯ go test -run="^$" -bench=. -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/pulumi/pulumi-aws/provider/v6
BenchmarkProvider-12                          76         153622535 ns/op
PASS
```

after:
```
❯ go test -run="^$" -bench="BenchmarkProvider$" -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/pulumi/pulumi-aws/provider/v6
BenchmarkProvider-12    	     150	  80806038 ns/op
PASS
```