@mjeffryes 2023-10-06
Learned on a different thread that we may be initializing the plugin even more than 4 times in some programs; given that context, I agree with @t0yv0 that it's probably worth shipping this at least as a temporary measure until we have a chance to more carefully separate runtime and gen-time concerns.